Remove reflection-based lightup S.R.M for netcoreapp
